BACKEND/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
sqlalchemy==2.0.40
werkzeug==2.3.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
qrcode==7.4.2
Pillow==10.1.0
pydantic-settings==2.0.3
email-validator==2.0.0 